Configured sql database

+++ app/src/main/resources/db/migration/V6__Create_invoice_entry.sql
@@ -5,7 +5,7 @@ CREATE TABLE public.invoice_entry
     price numeric(10, 2) NOT NULL DEFAULT 0,
     vat_value numeric(10, 2) NOT NULL DEFAULT 0,
     vat_rate bigint NOT NULL,
-    expense_related_to_car bigint NOT NULL,
+    expense_related_to_car numeric(10, 2) NOT NULL,
     PRIMARY KEY (id)
 );
